more optimizations after integration testing
