+++ .github/workflows/lint.yaml
@@ -1,0 +1,27 @@
+name: Lint
+
+on:
+  pull_request:
+    branches: [main]
+  # Also run on pushes to main so the badge stays accurate.
+  push:
+    branches: [main]
+
+permissions:
+  contents: read
+  # golangci-lint-action posts inline annotations on PR diffs.
+  pull-requests: read
+
+jobs:
+  golangci-lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-go@v5
+        with:
+          go-version: "1.24"
+      - name: golangci-lint
+        uses: golangci/golangci-lint-action@v6
+        with:
+          version: v2.8.0
+          args: --timeout=3m
