+++ massdriver/provisioning/artifacts/artifacts_test.go
@@ -1,4 +1,4 @@
-package resources_test
+package artifacts_test
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"github.com/go-resty/resty/v2"
 	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/internal/client"
 	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/internal/mockhttp"
-	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/services/resources"
+	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/provisioning/artifacts"
 	"github.com/stretchr/testify/require"
 )
 
@@ -25,7 +25,7 @@ func newTestClient(r *mockhttp.MockHTTPResponse) (*client.Client, *mockhttp.Muta
 	}, &roundtripper
 }
 
-func TestCreateResource(t *testing.T) {
+func TestCreateArtifact(t *testing.T) {
 	tests := []struct {
 		name         string
 		status       int
@@ -37,7 +37,7 @@ func TestCreateResource(t *testing.T) {
 		{
 			name:         "success",
 			status:       201,
-			sentBody:     `{"id":"abc-123","name":"Created","field":"database","type":"db","payload":{"foo":"bar","key":"value"}}`,
+			sentBody:     `{"name":"Created","type":"db","payload":{"foo":"bar","key":"value"}}`,
 			responseBody: `{"id":"abc-123","name":"Created"}`,
 			expectID:     "abc-123",
 			expectErr:    false,
@@ -50,20 +50,18 @@ func TestCreateResource(t *testing.T) {
 		},
 	}
 
-	input := resources.Resource{
-		ID:      "abc-123",
+	input := artifacts.Artifact{
 		Name:    "Created",
 		Type:    "db",
-		Field:   "database",
 		Payload: map[string]interface{}{"foo": "bar", "key": "value"},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			client, roundTripper := newTestClient(&mockhttp.MockHTTPResponse{StatusCode: tt.status, Body: tt.responseBody})
-			service := resources.NewService(client)
+			service := artifacts.NewService(client)
 
-			result, err := service.CreateResource(context.Background(), &input)
+			result, err := service.CreateArtifact(context.Background(), &input)
 
 			if tt.expectErr {
 				require.Error(t, err)
@@ -80,7 +78,7 @@ func TestCreateResource(t *testing.T) {
 	}
 }
 
-func TestGetResource(t *testing.T) {
+func TestGetArtifact(t *testing.T) {
 	tests := []struct {
 		name         string
 		status       int
@@ -116,9 +114,9 @@ func TestGetResource(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			client, _ := newTestClient(&mockhttp.MockHTTPResponse{StatusCode: tt.status, Body: tt.responseBody})
-			service := resources.NewService(client)
+			service := artifacts.NewService(client)
 
-			result, err := service.GetResource(context.Background(), "any-id")
+			result, err := service.GetArtifact(context.Background(), "any-id")
 
 			if tt.expectErr {
 				require.Error(t, err)
@@ -134,7 +132,7 @@ func TestGetResource(t *testing.T) {
 	}
 }
 
-func TestUpdateResource(t *testing.T) {
+func TestUpdateArtifact(t *testing.T) {
 	tests := []struct {
 		name         string
 		status       int
@@ -146,7 +144,7 @@ func TestUpdateResource(t *testing.T) {
 		{
 			name:         "success",
 			status:       200,
-			sentBody:     `{"id":"xyz-789","name":"Updated","field":"database","type":"db","payload":{"bar":"baz","x":"y"}}`,
+			sentBody:     `{"id":"xyz-789","name":"Updated","type":"db","payload":{"bar":"baz","x":"y"}}`,
 			responseBody: `{"id":"xyz-789","name":"Updated"}`,
 			expectErr:    false,
 			expectID:     "xyz-789",
@@ -159,10 +157,9 @@ func TestUpdateResource(t *testing.T) {
 		},
 	}
 
-	input := resources.Resource{
+	input := artifacts.Artifact{
 		ID:      "xyz-789",
 		Name:    "Updated",
-		Field:   "database",
 		Type:    "db",
 		Payload: map[string]interface{}{"bar": "baz", "x": "y"},
 	}
@@ -170,9 +167,9 @@ func TestUpdateResource(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			client, roundTripper := newTestClient(&mockhttp.MockHTTPResponse{StatusCode: tt.status, Body: tt.responseBody})
-			service := resources.NewService(client)
+			service := artifacts.NewService(client)
 
-			result, err := service.UpdateResource(context.Background(), "xyz-789", &input)
+			result, err := service.UpdateArtifact(context.Background(), "xyz-789", &input)
 
 			if tt.expectErr {
 				require.Error(t, err)
@@ -188,7 +185,7 @@ func TestUpdateResource(t *testing.T) {
 	}
 }
 
-func TestDeleteResource(t *testing.T) {
+func TestDeleteArtifact(t *testing.T) {
 	tests := []struct {
 		name         string
 		status       int
@@ -213,9 +210,9 @@ func TestDeleteResource(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			client, roundTripper := newTestClient(&mockhttp.MockHTTPResponse{StatusCode: tt.status, Body: tt.responseBody})
-			service := resources.NewService(client)
+			service := artifacts.NewService(client)
 
-			err := service.DeleteResource(context.Background(), "abc-123", "db")
+			err := service.DeleteArtifact(context.Background(), "abc-123", "db")
 
 			if tt.expectErr {
 				require.Error(t, err)
