+++ massdriver/provisioning/deployments/deployments_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-resty/resty/v2"
 	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/internal/client"
 	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/internal/mockhttp"
-	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/services/deployments"
+	"github.com/massdriver-cloud/massdriver-sdk-go/massdriver/provisioning/deployments"
 	"github.com/stretchr/testify/require"
 )
